+++ derive_builder_core/src/initializer.rs
@@ -2,6 +2,7 @@ use quote::{Tokens, ToTokens};
 use syn;
 use BuilderPattern;
 use Block;
+use Bindings;
 use DEFAULT_STRUCT_NAME;
 
 /// Initializer for the target struct fields, implementing `quote::ToTokens`.
@@ -46,8 +47,8 @@ pub struct Initializer<'a> {
     pub default_value: Option<Block>,
     /// Whether the build_method defines a default struct.
     pub use_default_struct: bool,
-    /// Whether the generated code should comply with `#![no_std]`.
-    pub no_std: bool,
+    /// Bindings to libstd or libcore.
+    pub bindings: Bindings,
 }
 
 impl<'a> ToTokens for Initializer<'a> {
@@ -81,7 +82,7 @@ impl<'a> Initializer<'a> {
         match self.builder_pattern {
             BuilderPattern::Owned => MatchSome::Move,
             BuilderPattern::Mutable |
-            BuilderPattern::Immutable => if self.no_std {
+            BuilderPattern::Immutable => if self.bindings.no_std {
                 MatchSome::CloneNoStd
             } else {
                 MatchSome::Clone
@@ -95,7 +96,7 @@ impl<'a> Initializer<'a> {
             Some(ref expr) => MatchNone::DefaultTo(expr),
             None => if self.use_default_struct {
                 MatchNone::UseDefaultStructField(self.field_ident)
-            } else if self.no_std {
+            } else if self.bindings.no_std {
                 MatchNone::ReturnErrorNoStd(format!("`{}` must be initialized", self.field_ident))
             } else {
                 MatchNone::ReturnError(format!("`{}` must be initialized", self.field_ident))
@@ -106,10 +107,9 @@ impl<'a> Initializer<'a> {
     fn default(&'a self) -> Tokens {
         match self.default_value {
             Some(ref expr) => quote!(#expr),
-            None => if self.no_std {
-                quote!(::core::default::Default::default())
-            } else {
-                quote!(::std::default::Default::default())
+            None => {
+                let default = self.bindings.default_trait();
+                quote!(#default::default())
             }
         }
     }
@@ -187,7 +187,7 @@ macro_rules! default_initializer {
             builder_pattern: BuilderPattern::Mutable,
             default_value: None,
             use_default_struct: false,
-            no_std: false,
+            bindings: Default::default(),
         }
     }
 }
@@ -281,7 +281,7 @@ mod tests {
     #[test]
     fn no_std() {
         let mut initializer = default_initializer!();
-        initializer.no_std = true;
+        initializer.bindings.no_std = true;
 
         assert_eq!(quote!(#initializer), quote!(
             foo: match self.foo {
@@ -296,7 +296,7 @@ mod tests {
     #[test]
     fn no_std_setter_disabled() {
         let mut initializer = default_initializer!();
-        initializer.no_std = true;
+        initializer.bindings.no_std = true;
         initializer.setter_enabled = false;
 
         assert_eq!(quote!(#initializer), quote!(